WD-1466 move blog notice date test to its own file
